pyproject.toml
[project]
name = "peach"
version = "0.0.0"
description = ""
authors = [{ name = "iamkenos", email = "alexis.matunog@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<=3.13"
dependencies = [
    "backoff (>=2.2.1,<3.0.0)",
    "behavex (>=4.6.0,<5.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-slugify[unidecode] (>=8.0.4,<9.0.0)",
    "inflection (>=0.5.1,<0.6.0)",
]
[dependency-groups]
dev = [
    "ruff (>=0.14.0,<0.15.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "peach"
packages = [{ include = "peach" }]

[tool.pyright]
reportMissingImports = "none"
reportUnknownMemberType = "none"
reportAttributeAccessIssue = "none"

[tool.ruff]
line-length = 150
target-version = "py312"
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"**/steps/__init__.py" = ["E402","F403"]
"**/steps/**/*.py" = ["F811"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

